feat(tabs): show the server connect widget on Library and Search when disconnected
Replace the logged-out error CTA on the Library and Search tabs with the full
JELLYFIN SERVER widget. The connect state machine moves from settings.tsx into
a reusable ServerConnectFlow (Settings behavior unchanged), ServerConnectScreen
hosts it full-screen, and useNetworkScan becomes a module-level shared store so
scan progress and results are identical across all three tabs and survive
screen unmounts.
